Added TranslatorZ3 to the Miasm test suite.
